requirements.txt
bcrypt==3.2.0
boto3==1.16.35
botocore==1.19.35
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1
jmespath==0.10.0
mongoengine==0.21.0
passlib==1.7.2
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pymongo==3.11.2
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.6
s3transfer==0.3.3
six==1.15.0
starlette==0.13.6
urllib3==1.26.2
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
requests